Removed AABBCollisionComponent. Characters have some collision built in (capsule) Split rendering out of AABBs, made AABBs immutable. Added AABB to Mesh, for the purpose of frustum culling. Added AABB scan method to BulletPhysicsRendering for checking for objects within an AABB

+++ src/main/java/org/terasology/model/structures/ViewFrustum.java
@@ -132,7 +132,7 @@ public class ViewFrustum {
      */
     public boolean intersects(AABB aabb) {
 
-        Vector3d[] aabbVertices = aabb.getVertices();
+        Vector3f[] aabbVertices = aabb.getVertices();
 
         Vector3f cp = CoreRegistry.get(WorldRenderer.class).getActiveCamera().getPosition();
